Normalize.css.
This is how I like doing submodules:

  git submodule add github.com/{user}/{repo} vendor/{repo}

Putting all referenced repos in the root vendor dir helps when projects have multiple different types of assets you want to reference (css + js). I then make a relative symlinks in the assets directory. This avoids external projects polluting dependencies. For instance, if a dependency has it's own jQuery, which one are we pulling in?